--- layers/emote-integrations/layers/integrations/betterttv/server/utils/api-types.ts
@@ -1,51 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export namespace BetterTTVApi {
-  interface BaseEmote {
-    animated: boolean;
-    code: string;
-    id: string;
-    imageType: "png" | "webp" | "gif";
-    height?: number;
-    width?: number;
-  }
-
-  export type GlobalEmote = Unwrap<
-    BaseEmote & {
-      modifier: boolean;
-      userId: string;
-    }
-  >;
-
-  type BaseUserEmote = Unwrap<
-    BaseEmote & {
-      codeOriginal?: string;
-    }
-  >;
-
-  export type ChannelEmote = Unwrap<
-    BaseUserEmote & {
-      userId: string;
-    }
-  >;
-
-  export type SharedEmote = Unwrap<
-    BaseUserEmote & {
-      user: {
-        displayName: string;
-        id: string;
-        name: string;
-        providerId: string;
-      };
-    }
-  >;
-
-  export type Emote = GlobalEmote | ChannelEmote | SharedEmote;
-
-  export interface User {
-    bots: string[];
-    channelEmotes: ChannelEmote[];
-    id: string;
-    avatar: string;
-    sharedEmotes: SharedEmote[];
-  }
-}
