+++ layers/emote-integrations/layers/integrations/betterttv/shared/types.global.ts
