+++ layers/emote-integrations/layers/integrations/frankerfacez/shared/types.person.ts
