+++ layers/emote-integrations/layers/integrations/frankerfacez/shared/types.ts
