+++ layers/emote-integrations/layers/integrations/seventv/shared/types.global.ts
