+++ layers/emote-integrations/layers/integrations/twitch/shared/types.global.ts
