+++ layers/emote-integrations/layers/integrations/twitch/shared/types.person.ts
